Move RiskAnalysisTempalte to Purpose Template Process package
