+++ packages/commons/src/risk-analysis/index.ts
@@ -1,7 +1,5 @@
 export * from "./models.js";
-export * from "./riskAnalysisValidationErrors.js";
-export * from "./riskAnalysisTemplateValidationErrors.js";
 export * from "./riskAnalysisTemplateErrors.js";
 export * from "./riskAnalysisValidation.js";
-export * from "./riskAnalysisTemplateValidation.js";
+export * from "./riskAnalysisValidationErrors.js";
 export * from "./rules/index.js";
